move type to types file
