+++ src/tabs/editor/components/single-company/EconomyDataTab.tsx
@@ -11,10 +11,8 @@ import type {
   GarboFieldMetadata,
   GarboReportingPeriodSummary,
 } from "../../lib/types";
-import {
-  updateReportingPeriods,
-  type ReportingPeriodWritePayload,
-} from "../../lib/companies-api";
+import { updateReportingPeriods } from "../../lib/companies-api";
+import type { ReportingPeriodWritePayload } from "../../lib/types";
 import { attachCompanyReportIdToPeriodPatch } from "../../lib/company-report-shells";
 import { inputClassName } from "../../lib/company-edit-utils";
 import {
