+++ src/tabs/editor/components/single-company/ReportingPeriodsDataTab.tsx
@@ -13,8 +13,8 @@ import type {
 import {
   deleteReportingPeriod,
   updateReportingPeriods,
-  type ReportingPeriodWritePayload,
 } from "../../lib/companies-api";
+import type { ReportingPeriodWritePayload } from "../../lib/types";
 import { attachCompanyReportIdToPeriodPatch } from "../../lib/company-report-shells";
 import { inputClassName } from "../../lib/company-edit-utils";
 import {
