+++ src/tabs/editor/lib/types.ts
@@ -88,6 +88,18 @@ export interface GarboReportingPeriodSummary {
   economy?: GarboEconomySummary;
 }
 
+/** One period in POST /api/companies/:id/reporting-periods. */
+export type ReportingPeriodWritePayload = {
+  startDate: string;
+  endDate: string;
+  companyReportId?: string;
+  reportURL?: string | null;
+  reportS3Url?: string | null;
+  reportSha256?: string | null;
+  emissions?: Record<string, unknown>;
+  economy?: Record<string, unknown>;
+};
+
 export interface GarboMinimalMetadata {
   user?: { name?: string | null } | null;
   verifiedBy?: { name: string } | null;
